+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.4] - 2026-02-12
+
+### Added
+
+- [#169] Add `use Koha::Plugins;` to scripts
+
+### Changed
+
+- [#170] Update mock_config to reflect changes in the plugin
+- [#171] Package update tries to delete a file that's not created 
+
+## [1.4.3] - 2027-02-09
+
+### Changed
+
+-[#164] Priority of holds should not be assigned
+
 ## [1.4.2] - 2026-02-05
 
 ### Added